movie_pjt_version2/backend/requirements.txt
Django==2.2.4
djangorestframework==3.10.2
joblib==0.13.2
numpy==1.17.0
pytz==2019.2
scikit-learn==0.21.3
scipy==1.3.0
sklearn==0.0
sqlparse==0.3.0
asn1crypto==0.24.0
cryptography==2.1.4
enum34==1.1.6
idna==2.6
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
pycrypto==2.6.1
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
